Don't change trailing slashes in paths, remove Net_URL dependency
Some ajax frameworks like Angularjs force you to use a trailing 
slash in your url. This is also a recommended habit by Google in 
order to avoid duplicate content. So it is now up to the route to 
define if there should be a trailing slash or not.

Net_URL_Mapper now ships with its own path resolve method, no more 
external dependency. The new method is a lot more robust.